file: TestDouble 폴더 생성 및 파일 분리 #6
- 파일 target 설정 변경
