+++ OpenMarket/OpenMarketTests/TestDouble/MockData.swift
@@ -1,0 +1,12 @@
+import UIKit
+
+struct MockData {
+    var healthData: Data {
+        let health = "\"OK\""
+        return health.data(using: .utf8)!
+    }
+    
+    var productListData: Data {
+        return NSDataAsset(name: "products")!.data
+    }
+}
