+++ OpenMarket/OpenMarketTests/TestDouble/MockURLSession.swift
@@ -1,24 +1,4 @@
-import UIKit
-@testable import OpenMarket
-
-struct MockData {
-    var healthData: Data {
-        let health = "\"OK\""
-        return health.data(using: .utf8)!
-    }
-    
-    var productListData: Data {
-        return NSDataAsset(name: "products")!.data
-    }
-}
-
-class MockURLSessionDataTask: URLSessionDataTask {
-    var resumeDidCall: () -> Void = {}
-
-    override func resume() {
-        resumeDidCall()
-    }
-}
+import Foundation
 
 class MockURLSession: URLSessionProtocol {
