+++ OpenMarket/OpenMarketTests/TestDouble/MockURLSessionDataTask.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+class MockURLSessionDataTask: URLSessionDataTask {
+    var resumeDidCall: () -> Void = {}
+
+    override func resume() {
+        resumeDidCall()
+    }
+}
